[Feat] #16 - 유저 정보로 가계부 제거
- 로그인 한 유저가 생성한 가계부를 조회하여 요청하는 파라미터와 일치하는 가계부 제거

+++ src/services/moneybook/moneybookService.js
@@ -43,19 +43,46 @@ const mbtiMoneybook = async (mbti) => {
   }
 };
 
-/**
- * @author 박성용
- * @version 1.0 22.7.6 최초 작성
- */
-const deleteMoneybook = async (moneybook_id) => {
-  try {
-    await Moneybook.destroy({
-      where: { id: moneybook_id },
-    });
-    return response(statusCode.OK, message.SUCCESS);
-  } catch (err) {
-    logger.error(`DB ERROR: ${err}`);
-    return errResponse(statusCode.DB_ERROR, message.DB_ERROR);
+const deleteMoneybook = async (userId, moneybook_id) => {
+  /**
+   * @author 박성용
+   * @version 1.0 22.7.6
+   * 가계부id 파라미터 요청시 해당 id의 가계부 제거
+   *
+   * @author 박성용
+   * @version 1.2 22.7.8
+   * 로그인한 유저 정보를 받아 (로그인한)유저가 생성한 가계부 내역 중 (가계부id)파라미터 요청과 일치하는 가계부만 제거
+   */
+
+  const userData = await Moneybook.findAll({
+    where: { user_id: userId },
+    attributes: ["id", "title", "user_id"],
+  });
+  let userDataList = [];
+  userData.forEach((data) => {
+    let setUserData = {
+      id: data.dataValues.id,
+      user_id: data.dataValues.user_id,
+      title: data.dataValues.title,
+    };
+    userDataList.push(setUserData);
+  });
+
+  for (let key in userDataList) {
+    let id = userDataList[key].id;
+
+    // 가계부의 id와 요청한 가계부 id 파라미터와 같다면
+    // 가계부 id에 해당하는 가계부를 삭제한다
+    if (id === parseInt(moneybook_id, 10)) {
+      await Moneybook.destroy({
+        where: { id: id },
+      });
+      return response(statusCode.OK, message.SUCCESS);
+    }
+    // 요청한 가계부id와 유저가 만든 가계부의 id가 일치하지 않는 경우
+    else if (moneybook_id !== id) {
+      return errResponse(statusCode.BAD_REQUEST, message.BAD_REQUEST);
+    }
   }
 };
 
@@ -63,7 +90,7 @@ const deleteMoneybook = async (moneybook_id) => {
  * @author 박성용
  * @version 1.0 22.7.6 최초 작성
  */
-const restoreMoneybook = async (moneybook_id) => {
+const restoreMoneybook = async (moneybook_id, userId) => {
   try {
     await Moneybook.restore({
       where: { id: moneybook_id },
